마커좌표설정(오류)

+++ main_map/app/src/main/java/com/example/jeyun/main_map/CalClass.java
@@ -47,7 +47,7 @@ public class CalClass {
         result.close();
     }
 
-    public int calMag(double input) {
+    public int calMag(int input) {
         HashMap<Integer, Double> hashtemp = new HashMap<>();
 
         Log.i("Calclass+DBsize", DBvalue.size() + "");
